Refactor 2 - Move node (#1027)
Summary:
Pull Request resolved: https://github.com/facebookresearch/spdl/pull/1027

Part 2 of https://github.com/facebookresearch/spdl/pull/1026

Moves `_node` into `_components`.

Differential Revision: D84854068
