+++ src/spdl/pipeline/_build.py
@@ -17,9 +17,9 @@ from typing import TypeVar
 
 from . import _config
 from ._components._hook import TaskHook
+from ._components._node import _build_pipeline_node, _run_pipeline_coroutines
 from ._components._pipe import _get_fail_counter
 from ._components._queue import AsyncQueue
-from ._node import _build_pipeline_node, _run_pipeline_coroutines
 from ._pipeline import Pipeline
 from .defs._defs import PipelineConfig
